fix: API Req 과정에서 올바른 타입의 값을 요청받도록 수정

+++ src/main/java/com/synergy/backend/domain/session/dto/sessionparticipateDto/SessionParticipateRateResDto.java
@@ -2,7 +2,7 @@ package com.synergy.backend.domain.session.dto.sessionparticipateDto;
 
 public record SessionParticipateRateResDto(
         String title,
-        Integer currentAttendee,
+        Long currentAttendee,
         Integer maximumAttendee
 ) {
 }
