+++ src/main/java/com/synergy/backend/domain/session/dto/sessionparticipateDto/SessionParticipateTechResDto.java
@@ -2,6 +2,6 @@ package com.synergy.backend.domain.session.dto.sessionparticipateDto;
 
 public record SessionParticipateTechResDto(
         String tech,
-        Integer attendeeCount
+        Long attendeeCount
 ) {
 }
